examples/inference/performance-showcase/requirements.txt
pillow
tensorflow
transformers
